requirements.txt
google-adk
pydantic>=2.0.0
python-dotenv
typing-extensions
reportlab>=4.0.0 